[release] ログイン時にサーバーに存在するエネミー情報を取得する処理の実装,他バグ修正

+++ DxGameClient/program/game/Actor/ActorData.h
@@ -51,8 +51,8 @@ public:
 		return level;
 	}
 
-	inline void  UpdateHp(float moveHp) {
-		HP += moveHp;
+	inline void  UpdateHp(double moveHp) {
+		HP += static_cast<float>(moveHp);
 	}
 
 	inline void SetHP(float startHP) {
